[HWKMETRICS-200] fix REST API test for counter rates
Now that the rates job is scheduled per tenant during tenant creation, the test
needs to explicitly create the tenant.

I also updated the createTenant method to concat the observable returned from
scheduling the task with the index updates. This way the subscriber will be
notified if any of the updates, along with scheduling the rates task, fails.