[HWKMETRICS-204] initial commit for task to create tenants
We query the metrics_idx table to determine whether or not a metric exists.
Metrics are implicitly created when inserting data points because we write to
both the data and the metrics_idx tables. We need to implicitly create tenants
in order to set up various background jobs like generating rates and deleting
data (see HWKMETRICS-191 for details).