[HWKMETRICS-204] fix schema and implicit tenant creation logic
The tenant column in tenants_by_time needs to be a clustering column.
CreateTenants has been updated to properly filter out id of tenants that
already exist. A new tenant creation method in MetricsServiceImpl is used. It
does not use a light weight transaction, nor does it update the retentions
index. There is no reason to incur the overhead of a LWT because its use is
intended for explicit tenant creation to ensure there is no naming conflict.