[HWKMETRICS-204] update tenants_by_time when inserting data, add REST API test
We now update the tenants_by_time table when inserting gauge data points. There
is also a REST API test to verify that the tenant is created when we only
insert gauge data points without explicitly creating the tenant. I will add
logic/tests for availability and counters in subsequent commits.

MetricsServiceLifecycle has also been updated to initialize scheduled jobs,
which now include the GenerateRate and CreateTenants classes. I do not like
having the job initialization logic, particularly the part of subscribing the
jobs, in MetricsServiceLifecyle. It should be a separate, core concern IMO.
This will probably be refactored in the near future once we see how the
scheduled jobs takes shape.